+++ core/datastore/src/commonMain/kotlin/com/mshdabiola/datastore/model/UserPreferences.kt
@@ -28,4 +28,6 @@ data class UserPreferences(
     val shouldHideOnboarding: Boolean = false,
     val shouldShowGradientBackground: Boolean = false,
     val language: String = "en-US",
+    val updateFromPreRelease: Boolean = true,
+    val showUpdateDialog: Boolean = true,
 )